Remove duplicate code + format
