+++ frontend/jupyter/src/app/pages/form/form-new/form-image-custom/form-image-custom.component.spec.ts
@@ -11,6 +11,8 @@ import { MatSelectModule } from '@angular/material/select';
 import { FormModule as KfFormModule } from 'kubeflow';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
 import { MatExpansionModule } from '@angular/material/expansion';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatCheckboxChange } from '@angular/material/checkbox';
 
 describe('FormImageCustomComponent', () => {
   let component: FormImageCustomComponent;
@@ -31,6 +33,7 @@ describe('FormImageCustomComponent', () => {
         ReactiveFormsModule,
         MatIconTestingModule,
         MatExpansionModule,
+        MatCheckboxModule,
       ],
     }).compileComponents();
   }));
@@ -41,6 +44,10 @@ describe('FormImageCustomComponent', () => {
     component.parentForm = new FormGroup({
       customImage: new FormControl(),
       customImageCheck: new FormControl(),
+      image: new FormControl(),
+      imageGroupOne: new FormControl(),
+      imageGroupTwo: new FormControl(),
+      imageGroupThree: new FormControl(),
     });
 
     fixture.detectChanges();
@@ -48,5 +55,28 @@ describe('FormImageCustomComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should toggle image controls when custom image is checked', () => {
+    const image = component.parentForm.get('image');
+    const imageGroupOne = component.parentForm.get('imageGroupOne');
+    const imageGroupTwo = component.parentForm.get('imageGroupTwo');
+    const imageGroupThree = component.parentForm.get('imageGroupThree');
+
+    let event = { checked: true } as MatCheckboxChange;
+    component.onSelect(event);
+
+    expect(image.disabled).toBe(true);
+    expect(imageGroupOne.disabled).toBe(true);
+    expect(imageGroupTwo.disabled).toBe(true);
+    expect(imageGroupThree.disabled).toBe(true);
+
+    event = { checked: false } as MatCheckboxChange;
+    component.onSelect(event);
+
+    expect(image.enabled).toBe(true);
+    expect(imageGroupOne.enabled).toBe(true);
+    expect(imageGroupTwo.enabled).toBe(true);
+    expect(imageGroupThree.enabled).toBe(true);
   });
 });
