+++ frontend/jupyter/src/app/pages/form/form-new/form-image/form-image.component.spec.ts
@@ -11,7 +11,6 @@ import { MatSelectModule } from '@angular/material/select';
 import { FormModule as KfFormModule } from 'kubeflow';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
 import { MatExpansionModule } from '@angular/material/expansion';
-import { MatCheckboxChange } from '@angular/material/checkbox';
 
 describe('FormImageComponent', () => {
   let component: FormImageComponent;
@@ -55,28 +54,5 @@ describe('FormImageComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
-  });
-
-  it('should toggle image controls when custom image is checked', () => {
-    const image = component.parentForm.get('image');
-    const imageGroupOne = component.parentForm.get('imageGroupOne');
-    const imageGroupTwo = component.parentForm.get('imageGroupTwo');
-    const imageGroupThree = component.parentForm.get('imageGroupThree');
-
-    let event = { checked: true } as MatCheckboxChange;
-    component.onSelect(event);
-
-    expect(image.disabled).toBe(true);
-    expect(imageGroupOne.disabled).toBe(true);
-    expect(imageGroupTwo.disabled).toBe(true);
-    expect(imageGroupThree.disabled).toBe(true);
-
-    event = { checked: false } as MatCheckboxChange;
-    component.onSelect(event);
-
-    expect(image.enabled).toBe(true);
-    expect(imageGroupOne.enabled).toBe(true);
-    expect(imageGroupTwo.enabled).toBe(true);
-    expect(imageGroupThree.enabled).toBe(true);
   });
 });
